+++ src/main/java/jeonseguard/backend/auth/domain/RefreshTokenStore.java
@@ -1,4 +1,4 @@
-package jeonseguard.backend.auth.domain.repository;
+package jeonseguard.backend.auth.domain;
 
 public interface RefreshTokenStore {
     void saveRefreshToken(Long userId, String refreshToken, long refreshExpirationTime);
